+++ SharpSqlBuilder/Maps/OrderMap.cs
@@ -8,52 +8,76 @@ namespace SharpSqlBuilder.Maps
 {
     public class OrderMap
     {
+        public readonly List<OrderMapItem> Items = new List<OrderMapItem>();
+
         private static readonly ConcurrentDictionary<Type, Dictionary<string, PropertyInfo>> OrderProps =
             new ConcurrentDictionary<Type, Dictionary<string, PropertyInfo>>();
 
-        public readonly Dictionary<string, OrderMapItem> Items;
-
-        public OrderMap()
+        
+        public OrderMap Append(SqlTable map, IEnumerable<NamedOrderItem> namedOrderItems,
+            bool throwOnMissingProperty = true)
         {
-        }
+            if (namedOrderItems == null)
+                throw new ArgumentException(nameof(namedOrderItems));
 
-        public OrderMap(params OrderMapItem[] items)
-        {
-            Items = items.ToDictionary(i => i.OrderProp);
-        }
-
-        private OrderMap(OrderMap one, OrderMap other)
-        {
-            Items = new Dictionary<string, OrderMapItem>(one.Items);
-            foreach (var item in other.Items)
+            var setOrders = namedOrderItems.Where(x => x != null).Select(p =>
             {
-                if (Items.ContainsKey(item.Key))
-                    throw new ArgumentException($"Key duplicate: {item.Key}");
+                var mapItem = map[p.PropertyName];
+                if (mapItem == null)
+                    if (throwOnMissingProperty)
+                        throw new ArgumentException($"Order property {p.PropertyName} have no counterpart in the map");
+                    else
+                        return null;
 
-                Items.Add(item.Key, item.Value);
-            }
+                var item = new OrderMapItem(mapItem, p.Direction);
+                return item;
+            }).Where(x => x != null);
+            Items.AddRange(setOrders);
+            return this;
         }
-
-
-        public static OrderMap FromOrder<TOrder>(SqlTable map) => FromOrder(map, typeof(TOrder));
-
-        public static OrderMap FromOrder(SqlTable map, Type orderType)
+        public OrderMap Append(OrderMap other)
         {
-            var orderProps = OrderProps.GetOrAdd(orderType, o => o.GetProperties().ToDictionary(p => p.Name));
+            return Append(other.Items);
+        }
+        public OrderMap Append(IEnumerable<OrderMapItem> orderItems)
+        {
+            Items.AddRange(orderItems);
+            return this;
+        }
+        public OrderMap Append(SqlTable map, object order)
+        {
+            if (order == null)
+                throw new ArgumentException(nameof(order));
+
+            var mapType = order.GetType();
+            var orderProps = OrderProps.GetOrAdd(mapType, o => o.GetProperties()
+                .Where(p => p.PropertyType == typeof(OrderItem)).ToDictionary(p => p.Name));
             var entityProps = map;
             var setOrders = orderProps.Values.Select(p =>
             {
                 var mapItem = map[p.Name];
                 if (mapItem == null)
                     throw new ArgumentException(
-                        $"Order property {p.Name} of type {orderType.Name} have no counterpart in the map");
+                        $"Order property {p.Name} of type {mapType.Name} have no counterpart in the map");
 
-                var item = new OrderMapItem(mapItem, p.Name);
-                return item;
-            });
-            return new OrderMap(setOrders.ToArray());
+
+                var orderItem = (OrderItem)p.GetValue(order);
+                return (mapItem, orderItem);
+            })
+                .Where(g=>g.orderItem!=null)
+                .OrderBy(g => g.orderItem.Index)
+                .Select(g => new OrderMapItem(g.mapItem, g.orderItem.Direction));
+            return Append(setOrders);
+        }
+        public static OrderMap Build(SqlTable map, IEnumerable<NamedOrderItem> namedOrderItems,
+            bool throwOnMissingProperty = true)
+        {
+            return new OrderMap().Append(map, namedOrderItems, throwOnMissingProperty);
         }
 
-        public OrderMap Concat(OrderMap other) => new OrderMap(this, other);
+        public static OrderMap FromOrder<T>(SqlTable map, T order) where T : class
+        {
+            return new OrderMap().Append(map, order);
+        }
     }
 }