+++ SharpSqlBuilder/NamedOrderItem.cs
@@ -1,0 +1,8 @@
+﻿namespace SharpSqlBuilder
+{
+    public class NamedOrderItem
+    {
+        public string PropertyName;
+        public OrderDirection Direction;
+    }
+}